+++ packages/cortex-react/docs/input.stories.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
 import { StoryFn } from '@storybook/react';
+import React from 'react';
 import { Input } from '../src';
 
 export default {