+++ packages/cortex-react/docs/select-grouped.stories.tsx
@@ -1,6 +1,7 @@
-import React, { useState } from 'react';
 import { StoryFn } from '@storybook/react';
+import React, { useState } from 'react';
 import { Select } from '../src';
+import SearchInput from '../src/components/SearchInput';
 
 export default {
   title: 'Cortex/Select/Grouped',
@@ -19,52 +20,63 @@ export default {
   },
 };
 
+const map = new Map([['São Paulo', [{ key: 'SP1', label: 'São Paulo' },
+  { key: 'SP2', label: 'Campinas' },
+  { key: 'SP3', label: 'Santos' }]],
+['Rio de Janeiro', [
+  { key: 'RJ1', label: 'Rio de Janeiro' },
+  { key: 'RJ2', label: 'Niterói' },
+  { key: 'RJ3', label: 'Petrópolis' },
+]],
+['Minas Gerais', [
+  { key: 'MG1', label: 'Belo Horizonte' },
+  { key: 'MG2', label: 'Uberlândia' },
+  { key: 'MG3', label: 'Ouro Preto' },
+]],
+['Bahia', [
+  { key: 'BA1', label: 'Salvador' },
+  { key: 'BA2', label: 'Feira de Santana' },
+  { key: 'BA3', label: 'Ilhéus' },
+]],
+['Paraná', [
+  { key: 'PR1', label: 'Curitiba' },
+  { key: 'PR2', label: 'Londrina' },
+  { key: 'PR3', label: 'Maringá' },
+]]
+])
+
 const Template: StoryFn = args => {
   const [value, setValue] = useState<{ key: string; label: string }>();
+  const [options, setOptions] = useState(map)
 
-  const estadosECidades = new Map();
-
-  estadosECidades.set('São Paulo', [
-    { key: 'SP1', label: 'São Paulo' },
-    { key: 'SP2', label: 'Campinas' },
-    { key: 'SP3', label: 'Santos' },
-  ]);
-  estadosECidades.set('Rio de Janeiro', [
-    { key: 'RJ1', label: 'Rio de Janeiro' },
-    { key: 'RJ2', label: 'Niterói' },
-    { key: 'RJ3', label: 'Petrópolis' },
-  ]);
-  estadosECidades.set('Minas Gerais', [
-    { key: 'MG1', label: 'Belo Horizonte' },
-    { key: 'MG2', label: 'Uberlândia' },
-    { key: 'MG3', label: 'Ouro Preto' },
-  ]);
-  estadosECidades.set('Bahia', [
-    { key: 'BA1', label: 'Salvador' },
-    { key: 'BA2', label: 'Feira de Santana' },
-    { key: 'BA3', label: 'Ilhéus' },
-  ]);
-  estadosECidades.set('Paraná', [
-    { key: 'PR1', label: 'Curitiba' },
-    { key: 'PR2', label: 'Londrina' },
-    { key: 'PR3', label: 'Maringá' },
-  ]);
+  const handleSearch = React.useCallback((searchArg: string) => {
+    const newMap = new Map()
+    Array.from(options.entries()).filter( ([key, items]) =>  {
+      const itemsFiltered = items.filter(item =>  {
+        return new RegExp(searchArg, 'ig').test(item.label)})
+        if(itemsFiltered.length)
+          newMap.set(key, itemsFiltered)
+    } ) 
+    setOptions(newMap)
+  }, []);
 
   return (
     <div className={'w-[350px]'}>
-      <Select
-        variant={args.variant}
-        label={args.label}
-        // onSearch={handleSearch}
-        placeholderSearchInput={args.placeholderSearchInput}
-        labelExtractor={op => op.label}
+      <Select.Root 
+        value={value} 
+        labelExtractor={op => op.label} 
         keyExtractor={op => op.key}
-        onSelect={setValue}
-        value={value}
-        grouped={true}
-        groupedLabelExtractor={labelGroup => labelGroup}
-        options={estadosECidades}
-      />
+      >
+        <Select.Trigger label={args.label} />
+        <Select.Popover>
+          <SearchInput onChange={handleSearch} placeholder={args.placeholderSearchInput} className='px-deca'/>
+          <Select.GroupedOptions
+            groupedLabelExtractor={labelGroup => labelGroup} 
+            options={options} 
+            onSelect={setValue}
+          />
+        </Select.Popover>
+      </Select.Root>
     </div>
   );
 };