+++ packages/cortex-react/src/components/Select/context.ts
@@ -1,0 +1,14 @@
+import { Dispatch, SetStateAction, createContext } from 'react';
+
+interface SelectContextProps<T> {
+  value?: T;
+  open?: boolean;
+  setOpen?: Dispatch<SetStateAction<boolean>>;
+  keyExtractor: (value: T) => string;
+  labelExtractor: (value: T) => string;
+}
+
+export const SelectContext = createContext<SelectContextProps<any>>({
+  keyExtractor: () => '',
+  labelExtractor: () => '',
+});