+++ packages/cortex-react/src/components/index.ts
@@ -15,7 +15,7 @@ export * from './Input';
 export * from './Modal';
 export * from './ProgressBar';
 export * from './RangeCalendar';
-export { default as Select } from './Select';
+export * from './Select';
 export * from './Skeleton';
 export * from './Table';
 export * from './Tag';